# openapi.yaml
openapi: 3.0.3
info:
  title: piYès API
  version: 1.0.0
  description: |
    API du service bancaire mobile piYès (Haïti). Base : `/api/v1`.

    Caractéristiques :
    - **Auth** : JWT bearer (24 h) + refresh token httpOnly (30 j). MFA par OTP (5 tentatives, TTL 15 min).
    - **Ledger** : tous les mouvements de fonds passent par un journal en partie double
      (fonctions Supabase `piyes_ledger_post`). Idempotence garantie par la clé
      `Idempotency-Key` (générée automatiquement si absente).
    - **Montants** : en **centimes de gourde HTG** (entiers positifs) dans les requêtes.
    - **Erreurs** : format uniforme `{ "error": { "message", "code" } }`.
    - **Temps réel** : les clients peuvent s'abonner aux changements Supabase
      (User, Transaction, Contact, Notification).

    Vitrine Open Banking : endpoints `/obp/v3.1.0/*` (voir tag `OpenBanking`).
  contact:
    name: piYès
    email: paiements@piyes.ht
servers:
  - url: http://localhost:3000/api/v1
    description: Dev local (backend)
  - url: http://localhost:3000
    description: Dev local (racine – pour OBP)
  - url: https://piyes-backend.vercel.app/api/v1
    description: Production (Vercel)
...
security:
  - bearerAuth: []
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Historique des transactions
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 50 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Liste des transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
  /contacts:
    get:
      tags: [Contacts]
      summary: Liste des contacts
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste des contacts
  /contacts/sync:
    post:
      tags: [Contacts]
      summary: Synchroniser les contacts du téléphone
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Contacts synchronisés
  /contacts/suggestions:
    get:
      tags: [Contacts]
      summary: Suggestions de contacts (utilisateurs piYès à découvrir)
      security: [{ bearerAuth: [] }]
      parameters:
        - {
            in: query,
            name: q,
            schema: { type: string },
            description: "Recherche par tag / nom / email / téléphone",
          }
        - { in: query, name: limit, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        tag: { type: string }
                        phone: { type: string }
                        email: { type: string }
                        avatarUrl: { type: string }
                        isUser: { type: boolean, default: true }
  /contacts/update/{id}:
    post:
      tags: [Contacts]
      summary: Mettre à jour un contact
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Contact mis à jour
  /contacts/{id}:
    patch:
      tags: [Contacts]
      summary: Modifier partiellement un contact
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Contact modifié
    delete:
      tags: [Contacts]
      summary: Supprimer un contact
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Contact supprimé

  # ------------------------------------------------------------ FRIENDSHIP
...
  /scheduler:
    get:
      tags: [Scheduler]
      summary: Liste des paiements planifiés
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Paiements planifiés
...
  /obp/v3.1.0/keys:
    post:
      tags: [OpenBanking]
      summary: Créer une clé API tiers (la clé brute n'est affichée qu'une fois)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, description: "Nom de la clé" }
      responses:
        "201":
          description: Clé créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey: { type: string }
                  id: { type: string }
                  name: { type: string }
                  status: { type: string }
    get:
      tags: [OpenBanking]
      summary: Lister les clés API (jamais les hashes)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste des clés
...
  /obp/v3.1.0/banks:
    get:
      tags: [OpenBanking]
      summary: Liste des banques (proxy OBP ou mock local)
      x-servers:
        - url: http://localhost:3000
        - url: https://piyes-backend.vercel.app
      security: [{ apiKeyAuth: [] }]
      responses:
        "200":
          description: Banques
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API tiers (Phase 4) — aussi acceptée via Bearer.
...
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: string }
        type:
          {
            type: string,
            enum:
              [
                DEPOSIT,
                WITHDRAW,
                TRANSFER,
                CARD_PAYMENT,
                INTERNATIONAL,
                RECHARGE,
                REQUEST,
                SCHEDULED,
                INTERBANK_OUT,
              ],
          }
        amount:
          { type: integer, description: "Centimes HTG (signé selon role)" }
        description: { type: string }
        date: { type: string, format: date-time }
        role: { type: string, enum: [PAYER, RECEIVER] }
        counterpartyName: { type: string }
        userId: { type: string }
        accountId: { type: string }
        external_id: { type: string }
...